Move WebSocket to Webserver and HttpJsonRpc is now async
