+++ libsrc/webconfig/QtHttpClientWrapper.h
@@ -9,6 +9,8 @@ class QTcpSocket;
 class QtHttpRequest;
 class QtHttpReply;
 class QtHttpServer;
+class WebSocketClient;
+class WebJsonRpc;
 
 class QtHttpClientWrapper : public QObject {
     Q_OBJECT
@@ -29,6 +31,7 @@ public:
     };
 
     QString getGuid (void);
+	void sendToClientWithReply (QtHttpReply * reply);
 
 private slots:
     void onClientDataReceived (void);
@@ -41,11 +44,13 @@ protected slots:
     void onReplySendDataRequested    (void);
 
 private:
-    QString         m_guid;
-    ParsingStatus   m_parsingStatus;
-    QTcpSocket    * m_sockClient;
-    QtHttpRequest * m_currentRequest;
-    QtHttpServer  * m_serverHandle;
+    QString           m_guid;
+    ParsingStatus     m_parsingStatus;
+    QTcpSocket    *   m_sockClient;
+    QtHttpRequest *   m_currentRequest;
+    QtHttpServer  *   m_serverHandle;
+	WebSocketClient * m_websocketClient = nullptr;
+	WebJsonRpc *      m_webJsonRpc = nullptr;
 };
 
 #endif // QTHTTPCLIENTWRAPPER_H
