+++ libsrc/webconfig/QtHttpHeader.cpp
@@ -3,30 +3,34 @@
 
 #include <QByteArray>
 
-const QByteArray & QtHttpHeader::Server             = QByteArrayLiteral ("Server");
-const QByteArray & QtHttpHeader::Date               = QByteArrayLiteral ("Date");
-const QByteArray & QtHttpHeader::Host               = QByteArrayLiteral ("Host");
-const QByteArray & QtHttpHeader::Accept             = QByteArrayLiteral ("Accept");
-const QByteArray & QtHttpHeader::Cookie             = QByteArrayLiteral ("Cookie");
-const QByteArray & QtHttpHeader::ContentType        = QByteArrayLiteral ("Content-Type");
-const QByteArray & QtHttpHeader::ContentLength      = QByteArrayLiteral ("Content-Length");
-const QByteArray & QtHttpHeader::Connection         = QByteArrayLiteral ("Connection");
-const QByteArray & QtHttpHeader::UserAgent          = QByteArrayLiteral ("User-Agent");
-const QByteArray & QtHttpHeader::AcceptCharset      = QByteArrayLiteral ("Accept-Charset");
-const QByteArray & QtHttpHeader::AcceptEncoding     = QByteArrayLiteral ("Accept-Encoding");
-const QByteArray & QtHttpHeader::AcceptLanguage     = QByteArrayLiteral ("Accept-Language");
-const QByteArray & QtHttpHeader::Authorization      = QByteArrayLiteral ("Authorization");
-const QByteArray & QtHttpHeader::CacheControl       = QByteArrayLiteral ("Cache-Control");
-const QByteArray & QtHttpHeader::ContentMD5         = QByteArrayLiteral ("Content-MD5");
-const QByteArray & QtHttpHeader::ProxyAuthorization = QByteArrayLiteral ("Proxy-Authorization");
-const QByteArray & QtHttpHeader::Range              = QByteArrayLiteral ("Range");
-const QByteArray & QtHttpHeader::ContentEncoding    = QByteArrayLiteral ("Content-Encoding");
-const QByteArray & QtHttpHeader::ContentLanguage    = QByteArrayLiteral ("Content-Language");
-const QByteArray & QtHttpHeader::ContentLocation    = QByteArrayLiteral ("Content-Location");
-const QByteArray & QtHttpHeader::ContentRange       = QByteArrayLiteral ("Content-Range");
-const QByteArray & QtHttpHeader::Expires            = QByteArrayLiteral ("Expires");
-const QByteArray & QtHttpHeader::LastModified       = QByteArrayLiteral ("Last-Modified");
-const QByteArray & QtHttpHeader::Location           = QByteArrayLiteral ("Location");
-const QByteArray & QtHttpHeader::SetCookie          = QByteArrayLiteral ("Set-Cookie");
-const QByteArray & QtHttpHeader::TransferEncoding   = QByteArrayLiteral ("Transfer-Encoding");
-const QByteArray & QtHttpHeader::ContentDisposition = QByteArrayLiteral ("Content-Disposition");
+const QByteArray & QtHttpHeader::Server               = QByteArrayLiteral ("Server");
+const QByteArray & QtHttpHeader::Date                 = QByteArrayLiteral ("Date");
+const QByteArray & QtHttpHeader::Host                 = QByteArrayLiteral ("Host");
+const QByteArray & QtHttpHeader::Accept               = QByteArrayLiteral ("Accept");
+const QByteArray & QtHttpHeader::Cookie               = QByteArrayLiteral ("Cookie");
+const QByteArray & QtHttpHeader::ContentType          = QByteArrayLiteral ("Content-Type");
+const QByteArray & QtHttpHeader::ContentLength        = QByteArrayLiteral ("Content-Length");
+const QByteArray & QtHttpHeader::Connection           = QByteArrayLiteral ("Connection");
+const QByteArray & QtHttpHeader::UserAgent            = QByteArrayLiteral ("User-Agent");
+const QByteArray & QtHttpHeader::AcceptCharset        = QByteArrayLiteral ("Accept-Charset");
+const QByteArray & QtHttpHeader::AcceptEncoding       = QByteArrayLiteral ("Accept-Encoding");
+const QByteArray & QtHttpHeader::AcceptLanguage       = QByteArrayLiteral ("Accept-Language");
+const QByteArray & QtHttpHeader::Authorization        = QByteArrayLiteral ("Authorization");
+const QByteArray & QtHttpHeader::CacheControl         = QByteArrayLiteral ("Cache-Control");
+const QByteArray & QtHttpHeader::ContentMD5           = QByteArrayLiteral ("Content-MD5");
+const QByteArray & QtHttpHeader::ProxyAuthorization   = QByteArrayLiteral ("Proxy-Authorization");
+const QByteArray & QtHttpHeader::Range                = QByteArrayLiteral ("Range");
+const QByteArray & QtHttpHeader::ContentEncoding      = QByteArrayLiteral ("Content-Encoding");
+const QByteArray & QtHttpHeader::ContentLanguage      = QByteArrayLiteral ("Content-Language");
+const QByteArray & QtHttpHeader::ContentLocation      = QByteArrayLiteral ("Content-Location");
+const QByteArray & QtHttpHeader::ContentRange         = QByteArrayLiteral ("Content-Range");
+const QByteArray & QtHttpHeader::Expires              = QByteArrayLiteral ("Expires");
+const QByteArray & QtHttpHeader::LastModified         = QByteArrayLiteral ("Last-Modified");
+const QByteArray & QtHttpHeader::Location             = QByteArrayLiteral ("Location");
+const QByteArray & QtHttpHeader::SetCookie            = QByteArrayLiteral ("Set-Cookie");
+const QByteArray & QtHttpHeader::TransferEncoding     = QByteArrayLiteral ("Transfer-Encoding");
+const QByteArray & QtHttpHeader::ContentDisposition   = QByteArrayLiteral ("Content-Disposition");
+const QByteArray & QtHttpHeader::Upgrade              = QByteArrayLiteral ("Upgrade");
+const QByteArray & QtHttpHeader::SecWebSocketKey      = QByteArrayLiteral ("Sec-WebSocket-Key");
+const QByteArray & QtHttpHeader::SecWebSocketProtocol = QByteArrayLiteral ("Sec-WebSocket-Protocol");
+const QByteArray & QtHttpHeader::SecWebSocketVersion  = QByteArrayLiteral ("Sec-WebSocket-Version");
